+++ ft-sdk/src/main/java/com/ft/sdk/garble/db/FTDBCachePolicy.java
@@ -137,6 +137,7 @@ public class FTDBCachePolicy {
      * @param optCount 写入数据数量
      */
     public void optLogCount(int optCount) {
+        if(enableLimitWithDbSize)return;
         logCount.addAndGet(optCount);
     }
 
@@ -146,6 +147,7 @@ public class FTDBCachePolicy {
      * @param optCount 写入数据数量
      */
     public void optRUMCount(int optCount) {
+        if(enableLimitWithDbSize)return;
         rumCount.addAndGet(optCount);
     }
 
@@ -252,10 +254,6 @@ public class FTDBCachePolicy {
 
     boolean isEnableLimitWithDbSize() {
         return enableLimitWithDbSize;
-    }
-
-    long getDbLimitSize() {
-        return dbLimitSize;
     }
 
     DBCacheDiscard getDbCacheDiscard() {